refactor(renderer): split reusable/ by subject; remove empty elements/
Step 4 of the renderer reorganization, final slice. The reusable/ folder —
named for a property rather than a subject — is dissolved:
- ui/tab-group.ts                    — a domain-free widget
- groups/elements/group-selector.ts, groups-for-applet.ts, groups-for-tool.ts
- self/profile/profile-detail.ts     — renders an agent's profile

elements/ is now empty and removed. The renderer is organized by
ontological subject: app/, self/, groups/, applets/, assets/, ui/.

Pure file moves + import-path updates; no behaviour change.
