+++ src/renderer/src/app/debugging-panel/state-dump.ts
@@ -23,7 +23,7 @@ import '@shoelace-style/shoelace/dist/components/button/button.js';
 import '../../groups/elements/group-context.js';
 import '../../applets/elements/applet-logo.js';
 import '../dialogs/create-group-dialog.js';
-import '../../elements/reusable/groups-for-applet.js';
+import '../../groups/elements/groups-for-applet.js';
 
 import { mossStyles } from '../../shared-styles.js';
 import { dateStr } from '../../utils.js';
