+++ src/renderer/src/assets/creatables/creatable-palette.ts
@@ -26,7 +26,7 @@ import '../pocket/wal-element.js';
 import '../pocket/pocket-search.js';
 import './creatable-view.js';
 import '../../app/navigation/group-applets-row.js';
-import '../../elements/reusable/group-selector.js';
+import '../../groups/elements/group-selector.js';
 
 import { StoreSubscriber } from '@holochain-open-dev/stores';
 import { mossStyles } from '../../shared-styles.js';
