+++ src/renderer/src/groups/elements/group-selector.ts
