+++ src/renderer/src/groups/elements/groups-for-applet.ts
