+++ src/renderer/src/self/profile/profile-detail.ts
