+++ src/renderer/src/ui/tab-group.ts
@@ -2,7 +2,7 @@ import { html, LitElement, css, TemplateResult } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
 import { localized } from '@lit/localize';
 
-import { mossStyles } from '../../shared-styles.js';
+import { mossStyles } from '../shared-styles.js';
 
 @localized()
 @customElement('tab-group')
